US 03.02.01 #13 Scaling of recipes has now been implemented along with some other minor cosmetic changes
Ingredient now has cost and count as a float, keep that in mind.

+++ app/src/main/java/com/example/lunchmunch/IngredientItemFragment.java
@@ -53,8 +53,8 @@ public class IngredientItemFragment extends DialogFragment implements AdapterVie
     private Location location;
     private String priceInput;
     private String amountInput;
-    private Integer price;
-    private Integer amount;
+    private Float price;
+    private Float amount;
     private Date expirationDate;
 
     private OnFragmentInteractionListener listener;
@@ -236,7 +236,7 @@ public class IngredientItemFragment extends DialogFragment implements AdapterVie
         if (priceInput.equals("")) {
             errMsg += "Enter a number for price, ";
         } else { // otherwise we have a valid int and we can parse
-            price = Integer.parseInt(priceInput);
+            price = Float.parseFloat(priceInput);
             //xml restricts entering neg number
             if (price == 0) {
                 errMsg += "Enter a positive number for price, ";
@@ -247,7 +247,7 @@ public class IngredientItemFragment extends DialogFragment implements AdapterVie
         if (amountInput.equals("")) {
             errMsg += "Enter a number for amount, ";
         } else { // otherwise we have a valid int and we can parse
-            amount = Integer.parseInt(amountInput);
+            amount = Float.parseFloat(amountInput);
             //xml restricts entering neg number
             if (amount == 0) {
                 errMsg += "Enter a positive number for amount, ";
